+++ docs/src/components/ComponentDoc/ComponentControls/ComponentControlsCopyLink.tsx
@@ -5,27 +5,43 @@ import { Icon, Menu } from 'semantic-ui-react'
 import ComponentControlsToolTip from './ComponentControlsToolTip'
 
 export default class ComponentControlsCopyLink extends Component<any, any> {
-  state: any = {}
-  mounted: boolean
+  private mounted: boolean
+  private readonly btnLabel = 'Direct link'
 
-  static propTypes = {
+  public static propTypes = {
     anchorName: PropTypes.string,
     onClick: PropTypes.func,
   }
 
-  shouldComponentUpdate(nextProps, nextState) {
+  public state: any = {}
+
+  public shouldComponentUpdate(nextProps, nextState) {
     return this.state.active !== nextState.active
   }
 
-  componentDidMount() {
+  public componentDidMount() {
     this.mounted = true
   }
 
-  componentWillUnmount() {
+  public componentWillUnmount() {
     this.mounted = false
   }
 
-  handleClick = e => {
+  public render() {
+    const { anchorName } = this.props
+    const { active } = this.state
+
+    return (
+      <ComponentControlsToolTip content={active ? ' Copied Link!' : this.btnLabel}>
+        <Menu.Item href={`#${anchorName}`} onClick={this.handleClick}>
+          <Icon color={active ? 'green' : 'grey'} fitted name="linkify" size="large" />
+          {this.btnLabel}
+        </Menu.Item>
+      </ComponentControlsToolTip>
+    )
+  }
+
+  private handleClick = e => {
     const { onClick } = this.props
 
     e.preventDefault()
@@ -35,18 +51,5 @@ export default class ComponentControlsCopyLink extends Component<any, any> {
     setTimeout(this.resetActive, 3000)
   }
 
-  resetActive = () => this.mounted && this.setState({ active: false })
-
-  render() {
-    const { anchorName } = this.props
-    const { active } = this.state
-
-    return (
-      <ComponentControlsToolTip content={active ? ' Copied Link!' : 'Direct link'}>
-        <Menu.Item href={`#${anchorName}`} onClick={this.handleClick}>
-          <Icon color={active ? 'green' : 'grey'} fitted name="linkify" size="large" />
-        </Menu.Item>
-      </ComponentControlsToolTip>
-    )
-  }
+  private resetActive = () => this.mounted && this.setState({ active: false })
 }