+++ docs/src/components/ComponentDoc/ComponentDoc.tsx
@@ -1,6 +1,6 @@
 import _ from 'lodash'
 import PropTypes from 'prop-types'
-import React, { Component } from 'react'
+import React, { Component, CSSProperties } from 'react'
 import DocumentTitle from 'react-document-title'
 import { withRouter } from 'react-router'
 import { Grid, Header, Icon } from 'semantic-ui-react'
@@ -16,7 +16,7 @@ import ComponentSidebar from './ComponentSidebar'
 import ComponentDocTag from './ComponentDocTag'
 
 const topRowStyle = { margin: '1em' }
-const exampleEndStyle = {
+const exampleEndStyle: CSSProperties = {
   textAlign: 'center',
   opacity: 0.5,
   paddingTop: '75vh',
@@ -103,11 +103,11 @@ class ComponentDoc extends Component<any, any> {
           </Grid.Row>
 
           <Grid.Row columns="equal">
-            <Grid.Column>
+            <Grid.Column style={{ padding: '0 0 0 28px' } as CSSProperties}>
               <div ref={this.handleExamplesRef}>
                 <ComponentExamples displayName={info.displayName} />
               </div>
-              <div style={exampleEndStyle as any}>
+              <div style={exampleEndStyle}>
                 This is the bottom <Icon name="pointing down" />
               </div>
             </Grid.Column>