+++ docs/src/components/Editor/index.tsx
@@ -1,0 +1,1 @@
+export { default, IEditorProps } from './Editor'